CRAS: use proper client_shm_fd var for error check
Code was recently added to check when streams are using a shm area that
is too large. However, it reads from a client_shm_fd variable that has
already been set to -1, so it triggers even for client shm streams.

Fix it to reference the stream_config instead.

BUG=None
TEST=open shm stream

Change-Id: Iaabbe0871ce6879a5573b2c2154c49c1e9876e7b
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/adhd/+/2040892
Tested-by: Fletcher Woodruff <fletcherw@chromium.org>
Reviewed-by: Chih-Yang Hsia <paulhsia@chromium.org>
Commit-Queue: Fletcher Woodruff <fletcherw@chromium.org>