CRAS: hfp_ag - Separate supported features of service record
Per HFP 1.7.1 spec, the values of the “SupportedFeatures” bitmap
given in Table 5.4 shall be the same as the values of the Bits 0
to 4 of the unsolicited result code +BRSF (see Section 4.34).

This change brings no functional change. As a preparation for
WBS feature, we're separating the features bit definition in
service record. Going forward this attribute value should be
controlled differently than in +BRSF command response.

BUG=b:149188379
TEST=None

Change-Id: Idf27586a159f2b0593141b81c94072777b7413c2
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/adhd/+/2073977
Reviewed-by: En-Shuo Hsu <enshuo@chromium.org>
Tested-by: Hsinyu Chao <hychao@chromium.org>
Commit-Queue: Hsinyu Chao <hychao@chromium.org>