CRAS: input_data - Skip SW capture gain only at AEC use case
For the case when external audio devices are selected as input
or output, which is not the typical AEC use case, CRAS shall
still take into account the SW capture gain.
This change aligns CRAS behavior to the case when APM runs inside
browser.

BUG=b:151794046
TEST=Manual test type-C headset.

Change-Id: I4784973c51c734d70821361d47c919d384010635
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/adhd/+/2108461
Reviewed-by: Chih-Yang Hsia <paulhsia@chromium.org>
Tested-by: Hsinyu Chao <hychao@chromium.org>
Commit-Queue: Hsinyu Chao <hychao@chromium.org>