CRAS: cras_alsa_io - Use the default gain on USB devices
In currect version, we set 20dB as the default gain of the USB device.
However, it is not accurate because the total gain user experienced is
the sum of microphone response and gain in USB chip.

We should trust the default gain USB headset manufacturer set when it
is shipped because it should work well with the microphone on it. In
addition, enable intrinsic sensitivity on USB and set the capture gain
to 0.

BUG=b:152173772
TEST=Verified the gain of USB devices did not change on grunt.

Change-Id: I1d97fa1fea40acc7cf58675922ceae6bfb2274c6
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/adhd/+/2116378
Reviewed-by: En-Shuo Hsu <enshuo@chromium.org>
Reviewed-by: Cheng-Yi Chiang <cychiang@chromium.org>
Tested-by: Yu-Hsuan Hsu <yuhsuan@chromium.org>
Commit-Queue: Yu-Hsuan Hsu <yuhsuan@chromium.org>