CRAS: Fix reply for AT+BIND and corresponding PTS
This CL fixes things that break the corresponding PTS
test(HFP/AG/SLC/BV-09-I).
1. We judged the command by the wrong position.
2. The PTS test somehow cannot parse the "BIND: (2)" response without
having 1 in it as "BIND: (1,2)", which is for "Enhanced Driver Status". Add it
and initiate it as disabled in the following "+BIND: 1,0". The test passed
after the change.

BUG=b:149188379
TEST=Run PTS tests

Change-Id: I5bf0e77564f6e8192c8ad3765dd4e41f0799ddc9
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/adhd/+/2126847
Reviewed-by: En-Shuo Hsu <enshuo@chromium.org>
Tested-by: En-Shuo Hsu <enshuo@chromium.org>
Commit-Queue: En-Shuo Hsu <enshuo@chromium.org>
Auto-Submit: En-Shuo Hsu <enshuo@chromium.org>