CRAS: Stop setting call status for HFP iodev
In chromium:270131, we set the call status when open/close the HFP iodev
to fix issues that some devices don't function when the callstatus is
not set. However, this breaks the qualification. As HFP/AG/ATA/BV-01-I
doesn't expect call status to be changed when the call is transferred
from HF to AG.

BUG=b:149188379, b:144738158
TEST=Run PTS, test on JBL flip 4, which also support hfp

Change-Id: I2d02563908fc6d97a7f5f4117a3c97a87d7e39bc
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/adhd/+/2160608
Tested-by: En-Shuo Hsu <enshuo@chromium.org>
Reviewed-by: Hsinyu Chao <hychao@chromium.org>
Commit-Queue: En-Shuo Hsu <enshuo@chromium.org>
Auto-Submit: En-Shuo Hsu <enshuo@chromium.org>