CRAS: Refine codec negotiation
Fix cases that we won't start codec negotiation process when HF supports
codec negotiation but supports only CVSD. Also replies ERROR if HF the
codec id in AT+BCS is invalid.

For codec id that is not consistent with the +BCS AG sent, there are known
cases that this happens because of delayed responses. For example, AG
sent +BCS:2 to HF but timeout to get HF's response. AG then sends a +BCS:1
as a fallback. At this timing, HF's response AT+BCS:2 for last +BCS:1
comes in, causing the inconsistent codec id AG sees. For this case, it
is inappropriate to regard the response HF sent as ERROR. We'll accept
it and rely on the following codec negotiation procedure to handle it.

This fixes the PTS tests:
HFP/AG/ACC/BV-09-I

BUG=chromium:980454, b:149188379, b:155282782
TEST=deploy and test QC35-ii and airpod2 can pair with NB and WB mode

Change-Id: I7096502f1804b097037d02ee29c6c273315fbaed
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/adhd/+/2172838
Reviewed-by: En-Shuo Hsu <enshuo@chromium.org>
Tested-by: En-Shuo Hsu <enshuo@chromium.org>
Commit-Queue: En-Shuo Hsu <enshuo@chromium.org>