CRAS: alsa_helpers: Fall back to max channels
Instead of giving up if none of the happy-list of channels is supported,
default to the max channel count of the card. This will allow cras to
function when new, high-channel-count devices are attached without
having to update the list manually.

For now this must be limited to 20 channels, until b/158509536 is fixed.

BUG=b:158506145
TEST=play to headphones of the TASCAM US-20x20.

Change-Id: I31fd3f9e845df88de3a76d302dd8de4eb1b9c5c0
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/adhd/+/2233793
Tested-by: Dylan Reid <dgreid@chromium.org>
Commit-Queue: Cheng-Yi Chiang <cychiang@chromium.org>
Reviewed-by: Cheng-Yi Chiang <cychiang@chromium.org>