libcras: allow reusing exact buffer indices
Generally, we prevent clients of libcras from setting the buffer offsets
of a buffer in the shm to overlap the other buffer offsets in the shm.
However, in the case that we are trying to set them to exactly the same
offsets, print a warning instead of reporting an error.

This is done because we occasionally trigger this path in crosvm's ac97
device because the intel8x0 kernel driver gives us the same buffer
offsets twice in a row. It is better to just have the audio glitch and
continue playback, than to break audio playback entirely.

BUG=chromium:1076885
TEST="aplay -D hw:0,0 -f dat /dev/urandom --period-size=256 --buffer-size=768"
     in the VM with CPU stress and check that audio doesn't stop.
TEST="aplay -D hw:0,0 -f dat /dev/urandom --period-size=256 --buffer-size=2816"
     in the VM with CPU stress and check that audio doesn't stop.

Change-Id: I0a436878b9eb9316d6ec660f846fb00bf02b43bc
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/adhd/+/2173871
Tested-by: Fletcher Woodruff <fletcherw@chromium.org>
Reviewed-by: Chih-Yang Hsia <paulhsia@chromium.org>
Commit-Queue: Fletcher Woodruff <fletcherw@chromium.org>