CRAS: Add UMA and a 5s delay before suspend
The suspend call is for handling cases that the Bluetooth device is
stuck and disconnects. The 5 second delay provides some space for a potential
recovery.

Add an UMA to track the frequency of different errors happened while
setting up the SCO connection to triage issues and validate our future fixes.

BUG=b:156886255
TEST=run the PTS test

Change-Id: Ife0361ac43b8d8002493575215babfb99b746362
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/adhd/+/2198851
Reviewed-by: Hsinyu Chao <hychao@chromium.org>
Tested-by: En-Shuo Hsu <enshuo@chromium.org>
Commit-Queue: En-Shuo Hsu <enshuo@chromium.org>