CRAS: bt_device - Prevent unwanted SCO socket error
When BT headset disconnects, it's a process consists of
multiple events. For example in a common case:

 1. CRAS got notified about A2DP disconnection.
 2. A2DP iodev destroyed.
 3. CRAS got notified about HFP disconnection.
 4. iodev destroyed.
 5. CRAS got notified about headset disconnection in baseband.

Between event 2 and 3 from above, CRAS tries to move active
playback streams to HFP iodev and fails at open iodev because
the SCO socket cannot be set up.
Although we know this error is no harm because the headset
disconnection is in progress, it may confuse others and make
it difficult to analyze the real SCO open failure.

This change prevents this failure by controlling the nodes' 'plug'
state by setting nodes to 'plugged' when supported profiles are
ready and to 'unplugged' at when the disconnection process begins.
Setting a node to 'unplugged' means it's no longer visible to UI
so audio thread leaves it alone for the rest of the disconnection
process.

BUG=b:156886255
TEST=Manual test with Bose QC35 II. Use the switch on headset to
power off during audio streaming, verify there's no 'SCO socket err'
printed in dmesg.

Change-Id: Iccaf662f72146709474414eb908afb342c383646
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/adhd/+/2209000
Reviewed-by: Hsinyu Chao <hychao@chromium.org>
Tested-by: Hsinyu Chao <hychao@chromium.org>
Commit-Queue: Hsinyu Chao <hychao@chromium.org>