CRAS: loopback_iodev: Fix sample_hook two-pass writing
There should be two pass of read to handle the case where write pointer
is near the end of buffer. Without this change, the first pass may only
read a part of the frames. Accumulating this effect will make loopback
stream starving of data and cause it to miss schedule.
This effects happens more severly when the block size of output stream
is large, because it will hit the end of buffer more frequently.

BUG=b:159416863, chromium:1091662
TEST=Use with the next patch. Play youtube.
     $cras_test_client --loopback_f /dev/null --block 480
     $cras_test_client --dump_a | less
      Check there is no RESCHEDULE in audio thread log
      Check there is no skipping in resulting audio recording.

Change-Id: I6f1f5ee661ce2a43cb69e7d1702e4c145a7734c3
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/adhd/+/2253682
Reviewed-by: Chih-Yang Hsia <paulhsia@chromium.org>
Tested-by: Cheng-Yi Chiang <cychiang@chromium.org>
Commit-Queue: Cheng-Yi Chiang <cychiang@chromium.org>