CRAS: hfp_alsa - Fix uninitialized channel layout in alsa iodev
hfp_alsa is implmented as a wrapper of the ALSA iodev associated
to the SCO PCM interface. There is a bug which leaves the
'channel_layout' of format uninitialized, and that indirectly
breaks the capture path by copying all zeros into shm.
To be specific, the value of ch_set in cras_audio_area_copy()
could be wrong.

BUG=b:157627735
TEST=Stress run bluetooth_AdapterAUSanity.au_hfp_nbs_dut_as_sink_test

Change-Id: Ie5abb6da8a6872f1a67411c115ac59fc3654ece5
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/adhd/+/2303416
Reviewed-by: Hsinyu Chao <hychao@chromium.org>
Tested-by: Hsinyu Chao <hychao@chromium.org>
Commit-Queue: Hsinyu Chao <hychao@chromium.org>