jack: enable all USB devices without jacks
If we have a device without a jack and multiple devices we should allow
all of the devices to be activated. In this case that we have multiple
nodes on a device (which means we have a UCM) then we should still
enable everything.

BUG=b:162852771
TEST=Plug in startech 7d card and verify all devices are plugged in

Change-Id: Iaeb136b60920c574d882b7051eaeefc7111c38ca
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/adhd/+/2340683
Tested-by: Curtis Malainey <cujomalainey@chromium.org>
Commit-Queue: Curtis Malainey <cujomalainey@chromium.org>
Reviewed-by: Yu-Hsuan Hsu <yuhsuan@chromium.org>
Auto-Submit: Curtis Malainey <cujomalainey@chromium.org>