CRAS: iodev - Error log instead of return at close failure
If iodev close fails, it should not just return because the
iodev's state would go wrong.

Note that this is nearly impossible to happen per current
implementation of all types of iodev. Fixing this can help
clarify if this is the cause to some crashes.

BUG=b:165134431
TEST=None

Change-Id: I9e9bbef14a6641877de9ea9809655aa2bb90a384
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/adhd/+/2397377
Reviewed-by: Cheng-Yi Chiang <cychiang@chromium.org>
Tested-by: Hsinyu Chao <hychao@chromium.org>
Commit-Queue: Hsinyu Chao <hychao@chromium.org>