CRAS: add codeql analysis
github does this scanning, might as well run it and see what it finds.
Factor out basic setup from the fuzzer Dockerfile to re-use it in the
github action.

This will only be run when someone manually pushes code to github. That
is still a pure win as we don't lose internal coverage and gain extra
insights.

BUG=none
TEST=push to github and check scan report. Build fuzzer and run locally.

Change-Id: I7a92a034aa39cd93143f8619f6b51f024bb5d820
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/adhd/+/2446462
Reviewed-by: Dylan Reid <dgreid@chromium.org>
Reviewed-by: Chih-Yang Hsia <paulhsia@chromium.org>
Commit-Queue: Dylan Reid <dgreid@chromium.org>
Tested-by: Dylan Reid <dgreid@chromium.org>