CRAS: alsa_jack - Non-HDMI/DP jack should not associate to ELD control
In find_gpio_jacks() we should check the device name to see if it's
really a HDMI/DP jack before associating an ELD mixer control.
This avoids issue on a problematic platform that a non-HDMI gpio
jack shares the same dev id with the ELD control of a real HDMI
device.

BUG=b:172928360
TEST=Verify 3.5mm jack detection works on hatch 13593.0.0

Change-Id: I10e607e975a9767cb0be958d267b8a985793bf2d
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/adhd/+/2534234
Commit-Queue: Hsinyu Chao <hychao@chromium.org>
Tested-by: Hsinyu Chao <hychao@chromium.org>
Reviewed-by: Yu-Hsuan Hsu <yuhsuan@chromium.org>