CRAS: device_monitor - Close erroneous device in main thread
If an output device hit I/O error in audio thread, there was
no explicit code path to close it. So if application keeps
the audio output stream alive, the output device would keep the
same buffer playing forever.

This change adds below handling logic:
When EIO is received, check if it's within 10 seconds from the
last EIO error. If not, then reset this device(close and reopen).
Otherwise treat this as a non-stop error and close this device
to avoid audio looping.

BUG=b:170100252
TEST=Use code to fake EIO error after 5 seconds of playback,
verify the device receives a reset, and 5 more seconds later
it gets closed gracefully.

Change-Id: I5e491ea6c23e550012cd65c9b4d51281b7047972
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/adhd/+/2681863
Reviewed-by: Curtis Malainey <cujomalainey@chromium.org>
Reviewed-by: Cheng-Yi Chiang <cychiang@chromium.org>
Commit-Queue: Hsinyu Chao <hychao@chromium.org>
Tested-by: Hsinyu Chao <hychao@chromium.org>