ucm: fix potential use after free bug
When a UCM has a suffix we create a name in alsa_card and pass it to
ucm_create which directly stores a reference to the string. When we
return from ucm_create we immediately free the string we passed leaving
a dangling pointer.

This has likely caused few issues as the name is rarely used except in a
few error cases.

BUG=None
TEST=emerge adhd

Change-Id: Icc24c4054b71e5f985a62c132d419c018bd56913
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/adhd/+/2601550
Reviewed-by: Chih-Yang Hsia <paulhsia@chromium.org>
Commit-Queue: Curtis Malainey <cujomalainey@chromium.org>
Tested-by: Curtis Malainey <cujomalainey@chromium.org>