Add support in CrAS for running NS and AGC
This CL adds the support in CrAS for extending the role of CrAS APM
to also  support for noise suppression (NS) and automatic gain control (AGC)
effects.

Functionality is added for flagging the availability of the NS and AGC effects
in board.ini, where the default value is for them to be off.

Furthermore, the CL adds support for allowing Chrome to enforce the AEC, NS and
AGC to be on regardless of the settings in board.ini. The purpose of this is to
prepare for any future finch-experimentation for launching CrAS APM, NS and AGC
on a broader scale without having to explicitly activate them in board.ini.

This CL must be landed after the related CL
-https://chromium-review.googlesource.com/c/chromiumos/third_party/webrtc-apm/+/2792582

BUG=b:177830918
TEST=Manually tested in calls on a Buddy and an Atlas device.

Cq-Depend: chromium:2792582
Change-Id: I8deb47a93d6b16df1543cf03ed936df1d66dca9c
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/adhd/+/2782881
Reviewed-by: Hsinyu Chao <hychao@chromium.org>
Tested-by: Hsinyu Chao <hychao@chromium.org>
Commit-Queue: Hsinyu Chao <hychao@chromium.org>