CRAS: empa_power - Restrict to S16_LE format
In the empa_power_init API we add new 'format' argument so
check against and disable the calculation if it's not
S16_LE.
This is to fix crash possibly because the audio devices runs
other formats. The calculation will be implemented for other
formats in future commits once we verify the crash no longer
shows up.

BUG=b:181740268
TEST=unittest

Change-Id: I21f569e6b88245e911bb402b8c1a860f613b65a3
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/adhd/+/2827562
Reviewed-by: Chih-Yang Hsia <paulhsia@chromium.org>
Commit-Queue: Hsinyu Chao <hychao@chromium.org>
Tested-by: Hsinyu Chao <hychao@chromium.org>