CRAS: Add GetNodeInfos method on dbus_control
In `GetNodes` method, the dbus signature of output nodes is `a{sv}`, but
the API returns multiple `a{sv}` objects, which is an undefined behavior
for chromeos-dbus-bingins. In chromeos-dbus-bindings, "an array of
dictionary" should be defined as `aa{sv}` and the generated interface
will be `std::vector<brillo::VariantDictionary>*`.

Since Chrome now depends on `GetNodes`, we need to add a new API
`GetNodeInfos` with signature `aa{sv}` for clients with
chromeos-dbus-bindings before deprecating the old one.

BUG=b:186192122
TEST=1. emerge-asurada adhd and cros deploy to DUT
     2. (DUT) restart cras
     3. (DUT) dbus-send --system --print-reply --dest=org.chromium.cras
     /org/chromium/cras org.chromium.cras.Control.GetNodeInfos
     4. Previous step shows the correct output

Change-Id: I8accaab8eb6d90eeb94594fad3477699c8f09767
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/adhd/+/2876318
Reviewed-by: Chih-Yang Hsia <paulhsia@chromium.org>
Tested-by: Yilin Yang (kerker) <kerker@chromium.org>
Commit-Queue: Yilin Yang (kerker) <kerker@chromium.org>