CRAS: Store Noise Cancellation support info in the node
In UCM case, the support of Noise Cancellation is determined by the
existence of UCM section modifier, which is defined with the node name.
Hence the support information should be a property stored in node
instance, as a bit flag of audio_effect.

In fill_node_list() of cras_iodev_list, the audio_effect property is
transfered to the list of cras_ionode_info which will be exposed by
GetNodes() on dbus interface.

BUG=b:170281884
TEST=Manual test

Change-Id: I403b53db14078ed1fc7ee8ea709a5c838408a422
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/adhd/+/2879118
Tested-by: Pin-chih Lin <johnylin@chromium.org>
Reviewed-by: Cheng-Yi Chiang <cychiang@chromium.org>
Commit-Queue: Pin-chih Lin <johnylin@chromium.org>
Auto-Submit: Pin-chih Lin <johnylin@chromium.org>