CRAS: Set correct volume control on Plantronics EncorePro 525 USB
There are two volume control on Plantronics EncorePro 525 USB. One is
'Sidetone' and another is 'Headset Earphone'. Only the second one works
but CRAS use the first one to control volume, which causes the volume
control not working.

Set 'Headset Earphone' as the volume control to fix this issue. In
addition, modify the capture gain to +5 dB gain on top of the default
capture gain according to the data reported by the headset.

BUG=b:175131804
TEST=volume control works while using Plantronics EncorePro 525 USB

Change-Id: Idfdbdeeb7d8d689e69afa5fb2aa27e986f3a0ef1
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/adhd/+/2879126
Commit-Queue: Yu-Hsuan Hsu <yuhsuan@chromium.org>
Tested-by: Yu-Hsuan Hsu <yuhsuan@chromium.org>
Reviewed-by: En-Shuo Hsu <enshuo@chromium.org>