cras-sys: don't recursively call default in default impl
The audio_dev_debug_info implementation of Default was changed to add a
call to Default::default for the same struct, which would result in a
recursive call and a stack overflow.  Explicitly initialize the new
member instead to avoid the problem.

The problem can be reproduced by a simple call to Default::default() for
the audio_dev_debug_info type, for example:

  let x: audio_dev_debug_info = Default::default();

This is also pointed out by clippy: "warning: function cannot return
without recursing".

BUG=None
TEST=cargo clippy

Change-Id: I3f939b0f3ae58a37c92c79b3f61edc0c9bbf27ce
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/adhd/+/2904247
Reviewed-by: Chih-Yang Hsia <paulhsia@chromium.org>
Tested-by: Chih-Yang Hsia <paulhsia@chromium.org>
Commit-Queue: Chih-Yang Hsia <paulhsia@chromium.org>