CRAS: Add a2dp_pcm_iodev and a2dp_manager
Add an a2dp pcm iodev that directly writing the PCM data to a socket
provided to it. This iodev will be used to connect with Floss.

The implementation is mostly borrowed from existing cras_a2dp_iodev.
The difference is that this pcm_iodev directly writes pcm data without
the encoding step in previous a2dp_iodev.

The cras_a2dp_manager is added for
- Manage the life cycle of a cras_a2dp_pcm_iodev
- Prepare the socket for the iodev to write PCM data to
- Keeping hooks for test tools

BUG=b:189276209
TEST=run cras and play audio on the mock device

Change-Id: I9e066c7920032f30028b36e30b400048f0ef7015
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/adhd/+/2959260
Tested-by: En-Shuo Hsu <enshuo@chromium.org>
Reviewed-by: Hsinyu Chao <hychao@chromium.org>
Commit-Queue: En-Shuo Hsu <enshuo@chromium.org>