cras-sys: in build.rs emit rerun-if command
According to
https://doc.rust-lang.org/cargo/reference/build-scripts.html#change-detection,
if the build script emits no change detection commands, the build script
is rerun every build, which will trigger rebuilds for everything that
transitively depends on cras-sys.

This change emits a change detection command for the main header bindgen
is wrapping. This has no effect on clean builds, but does cut down on
the time that incremental rebuilds will take. However, the headers
transitively included will not trigger the change detection because that
is only supported if the bindgen crate (instead of the binary) is being
used.

TEST=cargo build a dependent crate of cras-sys
BUG=None

Change-Id: I4cdaa9969fc393d5356ce2cdd3382deea7fd6dd8
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/adhd/+/2977151
Reviewed-by: Keiichi Watanabe <keiichiw@chromium.org>
Tested-by: Zach Reizner <zachr@chromium.org>
Commit-Queue: Zach Reizner <zachr@chromium.org>