CRAS: Only fallback to default_all_to_all when failed to remap CRAS_CH_FC
This CL changes the condition of fallback to the
default_all_to_all channel conversion.
This CL drops the input channels that cannot be re-mapped to the
allowed channel alternatives and only fallback to
default_all_to_all() when the CRAS_CH_FC fails to be
remapped. CRAS_CH_FC is not dropped as it usually contains
the dialogue content of audio.

This fixes the channel conversion inconsistency issue with
Chrome that users hear sometimes 2 channel audio and sometimes
{four, six} channel audio with the same stereo content on
{QUAD, six} channel devices.

BUG=b:181871882
TEST=1. Connect Scarlett 2i4
     2. `cras_test_client -c2 -P /tmp/sine.raw`
     3. `cras_test_client -c4 -P /tmp/sine.raw`
     4. Verify that only the front left and front right channel
        have audio.

Change-Id: Id43df88d0da66c4931962fcc20f3e9174f68c5a7
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/adhd/+/3066879
Tested-by: Judy Hsiao <judyhsiao@google.com>
Reviewed-by: Yu-Hsuan Hsu <yuhsuan@chromium.org>
Reviewed-by: Chih-Yang Hsia <paulhsia@chromium.org>
Commit-Queue: Judy Hsiao <judyhsiao@google.com>