cras: Fix a potential Divide-by-zero caught by fuzzer
normalize_factor = 1.0 / abs_sum;
If abs_sum = 0 then there will be a Divide-by-zero.

The abs_sum is the sum of a selected row of channel conversion
matrix, which should not be zero in normal cases.
Set the normalize_factor to 0 to prevent the Divide-by-zero
and print out an error message when `normalize` gets factor <= 0.

BUG=oss-fuzz:34195
TEST=`python infra/helper.py reproduce cras rclient_message
     ~/Downloads/clusterfuzz-testcase-minimized-rclient_message-6147736828248064`

Change-Id: Ic35396bb95f5bb49c20c4e50eb66ff4aa3ad4a08
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/adhd/+/3071503
Tested-by: Judy Hsiao <judyhsiao@google.com>
Reviewed-by: Chih-Yang Hsia <paulhsia@chromium.org>
Commit-Queue: Judy Hsiao <judyhsiao@google.com>