CRAS: apm_list - Configure APM capture process to mono
To work around an issue that APM might pick the 1st channel of
input, process and then writes to all output channels. This is
undesired because we want all channels, assuming HW provides
more than one, to be proccessed.

The exact condition to trigger this behavior is:
(1) More than one channel in APM input
(2) More than one channel in APM output
(3) multi_channel_capture is false

We're not ready to turn on multi_channel_capture so the best option
is to address (2). This is an acceptable fix because it makes APM's
behavior align with browser APM: the processed the output is always
mono, down-mixed of multiple channels from input.

BUG=b:190929775
TEST=Fake one channel to 0x00
cras_test_client -C /tmp/1 --effects aec
hexdump /tmp/1 to verify it's not full of 0x00

Cq-Depend: chromium:3067379
Change-Id: Idcc685b80c77c77134bb86851fb74ee3097c7e4b
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/adhd/+/3082081
Reviewed-by: Cheng-Yi Chiang <cychiang@chromium.org>
Tested-by: Hsinyu Chao <hychao@chromium.org>
Commit-Queue: Hsinyu Chao <hychao@chromium.org>