CRAS: bt_transport - Catch and reset abnormal write MTU values
CRAS uses the write MTU value returned from BlueZ to decide
how to encode one packet of A2DP audio data. We are seeing a
special case that this value is unreasonably high as 65535
and it breaks the data flow control.

This change work arounds the problem by setting a max
threshold to write mtu and cut it to A2DP_FIX_PACKET_SIZE.
Note that we don't know if this approach fixes all Bluetooth
audio devices with this problem. So we leave a warning log
for future tracking.

BUG=b:148222596
TEST=Connect Nest Hub and stream YouTube can work.

Change-Id: I20e408d2597af9520c1ff321d821510013e53238
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/adhd/+/3181241
Reviewed-by: En-Shuo Hsu <enshuo@chromium.org>
Commit-Queue: Hsinyu Chao <hychao@chromium.org>
Tested-by: Hsinyu Chao <hychao@chromium.org>