CRAS: Add cras_strerror wrapper
strerror is not thread safe and replaced with the cras_strerror
wrapper.

We use strerror in syslog for error messages.
In the extreme we might run into a race that one thread is writing
the strerror buffer while another thread is reading it.
However we only have two threads now so the described race is
highly unlikely.

BUG=b:201041751
TEST=USE=asan FEATURES=test emerge-octopus adhd

Change-Id: I048d2a62ac88daae3e4aa12fb801abd71be19a1b
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/adhd/+/3195278
Tested-by: Li-Yu Yu <aaronyu@google.com>
Reviewed-by: Chih-Yang Hsia <paulhsia@chromium.org>
Commit-Queue: Li-Yu Yu <aaronyu@google.com>