CRAS: bt_policy - Extract code into cras_bt_policy
cras_bt_device module acts as the place holder for Bluetooth
audio policy logics. Moving forward we will revisit some of the
policy logics. As a preparation for that, we are decoupling
these logics to a standalone, unittest-able module, and
eventually leave cras_bt_device as a module simply interacting
with BlueZ's device APIs.

To begin with, we're porting the profile switch logic and the
associated delay scheduling.

BUG=b:202505651
TEST=unittest

Change-Id: Idfea877a781c5e6ca59c0fcb8c36375cd594fb07
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/adhd/+/3213992
Tested-by: Hsinyu Chao <hychao@chromium.org>
Commit-Queue: Hsinyu Chao <hychao@chromium.org>
Reviewed-by: En-Shuo Hsu <enshuo@chromium.org>