cras_alsa_io - Fix the wrong max supported channels on a headphone
The CL:2518231 disabled the update of the max supported channels on a
headphone if it is unplugged. However, the value isn't updated when
the headphone is plugged. It causes the max supported channels on
headphone is always 0.

We can check the max supported channels of a headphone node at the
beginning since the value is fixed. In addition, always set the value to
2 for headphone nodes to save some time.

BUG=b:204134913
TEST=The value of max channels is correct on headphones.

Change-Id: I8b6bf1764df991e7bcdc1c929bfe9af60da2f94c
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/adhd/+/3244654
Reviewed-by: Judy Hsiao <judyhsiao@google.com>
Tested-by: Yu-Hsuan Hsu <yuhsuan@chromium.org>
Commit-Queue: Curtis Malainey <cujomalainey@chromium.org>