ucm: prevent double write of hotword model
This simply stops the double write we see from chrome, in reality we
should hunt down why this is happening and consolidate the logic.

It seems Chrome when the NC flag is set is setting the hotword model
multiple times. This checks to see if it is the same as the existing
model or if the existing model is already enabled and aborts the set to
avoid stressing SOF.

BUG=b:195087318
TEST=check if headset unplug causes any writes to amixer

Change-Id: Ia8e7c9ddb5a90ac7f85fdd9208b44f45a25420fd
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/adhd/+/3258093
Commit-Queue: Curtis Malainey <cujomalainey@chromium.org>
Commit-Queue: Ben Zhang <benzh@chromium.org>
Auto-Submit: Curtis Malainey <cujomalainey@chromium.org>
Tested-by: Curtis Malainey <cujomalainey@chromium.org>
Reviewed-by: Ben Zhang <benzh@chromium.org>