cras: Always set the volume
Remove the check of the device state and always set the volume even if
the device is not active. The previous behavior to ignore the event may
cause some USB headsets to regard the set volume event as failed and try
to revert the behavior by sending an extra event.

BUG=b:196969844, b:196184884
TEST=Verify the HW520 volume controls work without an active stream

Change-Id: Ibfd773cdfff5908ff1b29e6aba54bd00260d4ece
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/adhd/+/3282325
Tested-by: En-Shuo Hsu <enshuo@chromium.org>
Reviewed-by: Yu-Hsuan Hsu <yuhsuan@chromium.org>
Commit-Queue: En-Shuo Hsu <enshuo@chromium.org>