Add functionality in CrAS to control audio effects on the DSP
This CL adds functionality in CrAS that allows offloading the
AEC, NS and AGC functionalities from CrAS to DSP when the DSP
supports that.

The functionality relies on control from the browser stating
whether the offloading is allowed. That purpose of that scheme
is to allow finch experiments/flags to control the offloading
and allow fall-backs to the Browser audio effects.

The CL that allows the corresponding code in the Browser is
https://chromium-review.googlesource.com/c/chromium/src/+/3330293

This CL does not have any effect at all unless the board.ini contains
specific field specifying DSP support for AEC, NS, OR AGC, and at the
same time the browser finch flags for allowing those are specified.

The CL includes the changes in
https://chromium-review.googlesource.com/c/chromiumos/third_party/adhd/+/3334566
https://chromium-review.googlesource.com/c/chromiumos/third_party/adhd/+/3334567
which have been authored by cujomalainey@. Those CLs partly rewrite the UCM
handling.

BUG=b:191697965
TEST=Tested manually

Change-Id: Iab10a815949aaed67ae83aed32d540cce84316d3
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/adhd/+/3338425
Reviewed-by: Curtis Malainey <cujomalainey@chromium.org>
Tested-by: Per Åhgren <peah@chromium.org>
Commit-Queue: Per Åhgren <peah@chromium.org>