CRAS: Log the status of Noise Cancellation
It is used for tracing the user behavior on Noise Cancellation (NC).
When the device is opened, submit the status to the UMA server.

There are 3 cases as follows:
1) NC is supported and the UI toggle is shown and on --> NC enabled
2) NC is supported and the UI toggle is shown and off --> NC disabled
3) NC is blocked (by the selection of Internal Spkr node), the UI toggle
   is hidden --> NC disabled

Although NC is disabled in case2 and case3, the meaning is completely
different. Case2 indicates that users subjectively turn off NC through
the UI toggle, while for case3 NC is disabled due to the current use
case limitation.

BUG=b:170281884
TEST=Record audio and check the chrome://histograms

Change-Id: Iea2e8aaf8e22f492156f1a3288c8296506baa938
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/adhd/+/3353556
Reviewed-by: Yu-Hsuan Hsu <yuhsuan@chromium.org>
Commit-Queue: Pin-chih Lin <johnylin@chromium.org>
Tested-by: Pin-chih Lin <johnylin@chromium.org>
Auto-Submit: Pin-chih Lin <johnylin@chromium.org>