CRAS: rstream - Calculate ewma power properly on double buffer
When playback data outputs to iodev and rstream to mark data
as read, the chunk of data could sit across the double buffer
so it's not one continuous buffer.
ewma power calculation runs on continuous buffer, so to be safe
we should call it on each segment separately.
This CL fixes the SIGSEGV crash in ewma_power_calculate() when
the data to calculate is across the buffer boundary.

BUG=b:213539580,b:214150703,b:213525066
TEST=atest AudioRecordTest

Change-Id: Iadfe90ac48b9f63c575e341c8d371c78fe8b9006
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/adhd/+/3380121
Commit-Queue: Hsinyu Chao <hychao@chromium.org>
Reviewed-by: Judy Hsiao <judyhsiao@google.com>
Tested-by: Judy Hsiao <judyhsiao@google.com>
Reviewed-by: Chih-Yang Hsia <paulhsia@chromium.org>
Commit-Queue: Chih-Yang Hsia <paulhsia@chromium.org>
Tested-by: Chih-Yang Hsia <paulhsia@chromium.org>