CRAS: stream_apm - Add grace-period and fallback in the switching CRAS APM of stereo processing.
This CL changes the behavior in CRAS for detecting stereo content in the render
signal so that it does not start sending a stereo AEC reference to APM until
stereo  content has been observed for at least 2 seconds.
It also adds a behavior that falls-back to do mono processing if only mono
content has been obsererved for 5 minutes.

The background to this is to avoid short UI sounds that are stereo causes CRAS
APM to start processing in stereo.

BUG=chromium:1295710
TEST=Manually tested

Change-Id: I7c2927605ed5fa6b5be92935e4d561cc8976ac68
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/adhd/+/3455542
Reviewed-by: Cheng-Yi Chiang <cychiang@chromium.org>
Reviewed-by: Hsinyu Chao <hychao@chromium.org>
Tested-by: Per Åhgren <peah@google.com>
Commit-Queue: Per Åhgren <peah@chromium.org>