sound_card_init: fix infinite looping of PlayZeroWorker
The PlayZeroWorker may enter an infinite loop if the `thread_run` is set
to false in the PlayZeroWorkerInfo::destroy() and then set to true by
the PlayZeroWorker.
Fix the possibility of PlayZeroWorker to enter the infinite loop by:
1. Remove the thread_run initialization to true in the PlayZeroWorker.
2. Always exit the while loop after max_playback_iterations.

Also increase the TIMEOUT of waiting for PlayZeroWorker to start the
playback as it can take more than 1000 ms to get
CRAS_NODE_TYPE_INTERNAL_SPEAKER during the boot time.

BUG=b:219693872
TEST=reboot and verify that there is no sound_card_init playback
     stream after reboot.

Change-Id: I2de136395e5f1a3a03a8e3061b1329e520f2801c
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/adhd/+/3483049
Tested-by: Judy Hsiao <judyhsiao@google.com>
Auto-Submit: Judy Hsiao <judyhsiao@google.com>
Reviewed-by: Chih-Yang Hsia <paulhsia@chromium.org>
Commit-Queue: Curtis Malainey <cujomalainey@chromium.org>